integrate with dataproto

+++ src/m3/model/checker/GreaterChecker.java
@@ -1,10 +1,10 @@
 package m3.model.checker;
 
 
-public class GreaterChecker extends Checker<Double> {
+public class GreaterChecker extends Checker<Float> {
 
     @Override
-    public boolean check(Double o1, Double o2) {
+    public boolean check(Float o1, Float o2) {
         return o1 > o2;
     }
